backend/requirements.txt
asgiref==3.7.2
dill==0.3.7
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.5.1
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
setuptools==68.2.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
wheel==0.41.3
XlsxWriter==3.1.9
